requirements.txt
python-telegram-bot>=20.0
easyocr>=1.6.2
playwright>=1.35.0
python-dotenv>=1.0.0
opencv-python>=4.7.0
Pillow>=9.5.0
torch>=2.0.0
torchvision>=0.15.0
pymupdf>=1.22.0
